+++ 2023-rust/src/solutions/day4.rs
@@ -6,7 +6,7 @@ use crate::utils::load_data;
 ///! Code quality might not be optimal
 ///!
 
-fn get_correct_numbers(numbers: &(Vec<u32>, Vec<u32>)) -> u32 {
+fn get_correct_numbers(numbers: &(Vec<i64>, Vec<i64>)) -> i64 {
     let mut c = 0;
     for holding in &numbers.1 {
         if numbers.0.contains(holding) {
@@ -16,14 +16,14 @@ fn get_correct_numbers(numbers: &(Vec<u32>, Vec<u32>)) -> u32 {
     c
 }
 
-fn get_points(correct_numbers: u32) -> u32 {
+fn get_points(correct_numbers: i64) -> i64 {
     if correct_numbers <= 1 {
         return correct_numbers;
     }
-    return 2u32.pow(correct_numbers - 1)
+    return 2i64.pow(correct_numbers as u32 - 1)
 }
 
-fn get_scratchcards(t: &(Vec<u32>, Vec<u32>), i: usize, d: &Vec<(Vec<u32>, Vec<u32>)>) -> u32 {
+fn get_scratchcards(t: &(Vec<i64>, Vec<i64>), i: usize, d: &Vec<(Vec<i64>, Vec<i64>)>) -> i64 {
     let mut c = get_correct_numbers(t);
     for idx in i + 1..i + usize::try_from(c).unwrap() + 1 {
         c += get_scratchcards(&d[idx], idx, d)
@@ -31,21 +31,21 @@ fn get_scratchcards(t: &(Vec<u32>, Vec<u32>), i: usize, d: &Vec<(Vec<u32>, Vec<u
     return c;
 }
 
-pub fn run(test: bool) -> (u32, u32) {
+pub fn run(test: bool) -> (i64, i64) {
     let data = load_data(4, test);
-    let data: Vec<(Vec<u32>, Vec<u32>)> = data.split('\n').map(|x| {
+    let data: Vec<(Vec<i64>, Vec<i64>)> = data.split('\n').map(|x| {
         let numbers = &x[x.chars().position(|c| c == ':').unwrap() + 1..];
         let numbers: Vec<&str> = numbers
             .split('|')
             .map(|y| y.trim())
             .collect();
-        let winning: Vec<u32> = numbers[0]
+        let winning: Vec<i64> = numbers[0]
             .split_whitespace()
-            .map(|y| y.parse::<u32>().unwrap())
+            .map(|y| y.parse::<i64>().unwrap())
             .collect();
-        let holding: Vec<u32> = numbers[1]
+        let holding: Vec<i64> = numbers[1]
             .split_whitespace()
-            .map(|y| y.parse::<u32>().unwrap())
+            .map(|y| y.parse::<i64>().unwrap())
             .collect();
         (winning, holding)
     }).collect();
